Use target_* functions to set compiler features
This is more modern CMake, and makes it easier to include Bifrost
in other CMake projects (as library for example).

One of the big benefits of the new target_* API is that compiler
options like -march=native now get propogated to any executable
linking to bifrost_static of bifrost_dynamic.